+++ proxy/proxy.py
@@ -8,21 +8,120 @@
     :copyright: (c) 2013-present by Abhinav Singh and contributors.
     :license: BSD, see LICENSE for more details.
 """
+import abc
+import argparse
+import base64
+import collections
 import contextlib
+import ipaddress
 import os
 import sys
 import time
 import logging
+import importlib
+import inspect
 
 from types import TracebackType
-from typing import List, Optional, Generator, Any, Type
+from typing import Dict, List, Optional, Generator, Any, Tuple, Type, Union, cast
 
-from .common.utils import bytes_
+from .common.utils import bytes_, text_
 from .common.flags import Flags
+from .common.types import IpAddress
+from .common.version import __version__
 from .core.acceptor import AcceptorPool
 from .http.handler import HttpProtocolHandler
+from .common.flag import flags
+from .common.constants import COMMA, DEFAULT_BASIC_AUTH, DEFAULT_DISABLE_HTTP_PROXY, PY2_DEPRECATION_MESSAGE
+from .common.constants import DEFAULT_ENABLE_DASHBOARD, DEFAULT_ENABLE_DEVTOOLS
+from .common.constants import DEFAULT_ENABLE_STATIC_SERVER, DEFAULT_ENABLE_WEB_SERVER
+from .common.constants import DEFAULT_LOG_FILE, DEFAULT_LOG_FORMAT, DEFAULT_LOG_LEVEL
+from .common.constants import DEFAULT_OPEN_FILE_LIMIT, DEFAULT_PID_FILE, DEFAULT_PLUGINS
+from .common.constants import DEFAULT_VERSION, DOT, PLUGIN_DASHBOARD, PLUGIN_DEVTOOLS_PROTOCOL
+from .common.constants import PLUGIN_HTTP_PROXY, PLUGIN_INSPECT_TRAFFIC, PLUGIN_PAC_FILE
+from .common.constants import PLUGIN_WEB_SERVER
+
+if os.name != 'nt':
+    import resource
 
 logger = logging.getLogger(__name__)
+
+
+flags.add_argument(
+    '--pid-file',
+    type=str,
+    default=DEFAULT_PID_FILE,
+    help='Default: None. Save parent process ID to a file.')
+flags.add_argument(
+    '--basic-auth',
+    type=str,
+    default=DEFAULT_BASIC_AUTH,
+    help='Default: No authentication. Specify colon separated user:password '
+    'to enable basic authentication.')
+flags.add_argument(
+    '--version',
+    '-v',
+    action='store_true',
+    default=DEFAULT_VERSION,
+    help='Prints proxy.py version.')
+flags.add_argument(
+    '--disable-http-proxy',
+    action='store_true',
+    default=DEFAULT_DISABLE_HTTP_PROXY,
+    help='Default: False.  Whether to disable proxy.HttpProxyPlugin.')
+flags.add_argument(
+    '--enable-dashboard',
+    action='store_true',
+    default=DEFAULT_ENABLE_DASHBOARD,
+    help='Default: False.  Enables proxy.py dashboard.'
+)
+flags.add_argument(
+    '--enable-devtools',
+    action='store_true',
+    default=DEFAULT_ENABLE_DEVTOOLS,
+    help='Default: False.  Enables integration with Chrome Devtool Frontend. Also see --devtools-ws-path.'
+)
+flags.add_argument(
+    '--enable-static-server',
+    action='store_true',
+    default=DEFAULT_ENABLE_STATIC_SERVER,
+    help='Default: False.  Enable inbuilt static file server. '
+    'Optionally, also use --static-server-dir to serve static content '
+    'from custom directory.  By default, static file server serves '
+    'out of installed proxy.py python module folder.'
+)
+flags.add_argument(
+    '--enable-web-server',
+    action='store_true',
+    default=DEFAULT_ENABLE_WEB_SERVER,
+    help='Default: False.  Whether to enable proxy.HttpWebServerPlugin.')
+flags.add_argument(
+    '--log-level',
+    type=str,
+    default=DEFAULT_LOG_LEVEL,
+    help='Valid options: DEBUG, INFO (default), WARNING, ERROR, CRITICAL. '
+    'Both upper and lowercase values are allowed. '
+    'You may also simply use the leading character e.g. --log-level d')
+flags.add_argument(
+    '--log-file',
+    type=str,
+    default=DEFAULT_LOG_FILE,
+    help='Default: sys.stdout. Log file destination.')
+flags.add_argument(
+    '--log-format',
+    type=str,
+    default=DEFAULT_LOG_FORMAT,
+    help='Log format for Python logger.')
+flags.add_argument(
+    '--open-file-limit',
+    type=int,
+    default=DEFAULT_OPEN_FILE_LIMIT,
+    help='Default: 1024. Maximum number of files (TCP connections) '
+    'that proxy.py can open concurrently.')
+flags.add_argument(
+    '--plugins',
+    type=str,
+    default=DEFAULT_PLUGINS,
+    help='Comma separated plugins')
 
 
 class Proxy:
@@ -33,7 +132,7 @@ class Proxy:
     """
 
     def __init__(self, input_args: Optional[List[str]], **opts: Any) -> None:
-        self.flags = Flags.initialize(input_args, **opts)
+        self.flags = Proxy.initialize(input_args, **opts)
         self.acceptors: Optional[AcceptorPool] = None
 
     def write_pid_file(self) -> None:
@@ -62,6 +161,226 @@ class Proxy:
         assert self.acceptors
         self.acceptors.shutdown()
         self.delete_pid_file()
+
+    @staticmethod
+    def initialize(input_args: Optional[List[str]], **opts: Any) -> Flags:
+        if not Proxy.is_py3():
+            print(PY2_DEPRECATION_MESSAGE)
+            sys.exit(1)
+
+        args = flags.parse_args(input_args)
+
+        # Print version and exit
+        if args.version:
+            print(__version__)
+            sys.exit(0)
+
+        # Setup logging module
+        Proxy.setup_logger(args.log_file, args.log_level, args.log_format)
+
+        # Setup limits
+        Proxy.set_open_file_limit(args.open_file_limit)
+
+        # Load plugins
+        default_plugins = Proxy.get_default_plugins(args)
+
+        # Load default plugins along with user provided --plugins
+        plugins = Proxy.load_plugins(
+            [bytes_(p) for p in collections.OrderedDict(default_plugins).keys()] +
+            [p if isinstance(p, type) else bytes_(p) for p in opts.get(
+                'plugins', args.plugins.split(text_(COMMA)))]
+        )
+
+        # proxy.py currently cannot serve over HTTPS and perform TLS interception
+        # at the same time.  Check if user is trying to enable both feature
+        # at the same time.
+        if (args.cert_file and args.key_file) and \
+                (args.ca_key_file and args.ca_cert_file and args.ca_signing_key_file):
+            print('You can either enable end-to-end encryption OR TLS interception,'
+                  'not both together.')
+            sys.exit(1)
+
+        # Generate auth_code required for basic authentication if enabled
+        auth_code = None
+        if args.basic_auth:
+            auth_code = b'Basic %s' % base64.b64encode(bytes_(args.basic_auth))
+
+        return Flags(
+            plugins=plugins,
+            auth_code=cast(Optional[bytes], opts.get('auth_code', auth_code)),
+            server_recvbuf_size=cast(
+                int,
+                opts.get(
+                    'server_recvbuf_size',
+                    args.server_recvbuf_size)),
+            client_recvbuf_size=cast(
+                int,
+                opts.get(
+                    'client_recvbuf_size',
+                    args.client_recvbuf_size)),
+            pac_file=cast(
+                Optional[str], opts.get(
+                    'pac_file', bytes_(
+                        args.pac_file))),
+            pac_file_url_path=cast(
+                Optional[bytes], opts.get(
+                    'pac_file_url_path', bytes_(
+                        args.pac_file_url_path))),
+            disable_headers=cast(Optional[List[bytes]], opts.get('disable_headers', [
+                header.lower() for header in bytes_(
+                    args.disable_headers).split(COMMA) if header.strip() != b''])),
+            certfile=cast(
+                Optional[str], opts.get(
+                    'cert_file', args.cert_file)),
+            keyfile=cast(Optional[str], opts.get('key_file', args.key_file)),
+            ca_cert_dir=cast(
+                Optional[str], opts.get(
+                    'ca_cert_dir', args.ca_cert_dir)),
+            ca_key_file=cast(
+                Optional[str], opts.get(
+                    'ca_key_file', args.ca_key_file)),
+            ca_cert_file=cast(
+                Optional[str], opts.get(
+                    'ca_cert_file', args.ca_cert_file)),
+            ca_signing_key_file=cast(
+                Optional[str],
+                opts.get(
+                    'ca_signing_key_file',
+                    args.ca_signing_key_file)),
+            ca_file=cast(
+                Optional[str],
+                opts.get(
+                    'ca_file',
+                    args.ca_file)),
+            hostname=cast(IpAddress,
+                          opts.get('hostname', ipaddress.ip_address(args.hostname))),
+            port=cast(int, opts.get('port', args.port)),
+            backlog=cast(int, opts.get('backlog', args.backlog)),
+            num_workers=cast(int, opts.get('num_workers', args.num_workers)),
+            static_server_dir=cast(
+                str,
+                opts.get(
+                    'static_server_dir',
+                    args.static_server_dir)),
+            enable_static_server=cast(
+                bool,
+                opts.get(
+                    'enable_static_server',
+                    args.enable_static_server)),
+            devtools_ws_path=cast(
+                bytes,
+                opts.get(
+                    'devtools_ws_path',
+                    getattr(args, 'devtools_ws_path', None))),
+            timeout=cast(int, opts.get('timeout', args.timeout)),
+            threadless=cast(bool, opts.get('threadless', args.threadless)),
+            enable_events=cast(
+                bool,
+                opts.get(
+                    'enable_events',
+                    args.enable_events)),
+            pid_file=cast(Optional[str], opts.get('pid_file', args.pid_file))
+        )
+
+    @staticmethod
+    def load_plugins(plugins: List[Union[bytes, type]]
+                     ) -> Dict[bytes, List[type]]:
+        """Accepts a comma separated list of Python modules and returns
+        a list of respective Python classes."""
+        p: Dict[bytes, List[type]] = {
+            b'HttpProtocolHandlerPlugin': [],
+            b'HttpProxyBasePlugin': [],
+            b'HttpWebServerBasePlugin': [],
+            b'ProxyDashboardWebsocketPlugin': []
+        }
+        for plugin_ in plugins:
+            if isinstance(plugin_, type):
+                module_name = '__main__'
+                klass = plugin_
+            else:
+                plugin = text_(plugin_.strip())
+                if plugin == '':
+                    continue
+                module_name, klass_name = plugin.rsplit(text_(DOT), 1)
+                klass = getattr(
+                    importlib.import_module(
+                        module_name.replace(
+                            os.path.sep, text_(DOT))),
+                    klass_name)
+            mro = list(inspect.getmro(klass))
+            mro.reverse()
+            iterator = iter(mro)
+            while next(iterator) is not abc.ABC:
+                pass
+            base_klass = next(iterator)
+            if klass not in p[bytes_(base_klass.__name__)]:
+                p[bytes_(base_klass.__name__)].append(klass)
+            logger.info('Loaded plugin %s.%s', module_name, klass.__name__)
+        return p
+
+    @staticmethod
+    def get_default_plugins(
+            args: argparse.Namespace) -> List[Tuple[str, bool]]:
+        # Prepare list of plugins to load based upon --enable-* and --disable-*
+        # flags
+        default_plugins: List[Tuple[str, bool]] = []
+        if args.enable_dashboard:
+            default_plugins.append((PLUGIN_WEB_SERVER, True))
+            args.enable_static_server = True
+            default_plugins.append((PLUGIN_DASHBOARD, True))
+            default_plugins.append((PLUGIN_INSPECT_TRAFFIC, True))
+            args.enable_events = True
+            args.enable_devtools = True
+        if args.enable_devtools:
+            default_plugins.append((PLUGIN_DEVTOOLS_PROTOCOL, True))
+            default_plugins.append((PLUGIN_WEB_SERVER, True))
+        if not args.disable_http_proxy:
+            default_plugins.append((PLUGIN_HTTP_PROXY, True))
+        if args.enable_web_server or \
+                args.pac_file is not None or \
+                args.enable_static_server:
+            default_plugins.append((PLUGIN_WEB_SERVER, True))
+        if args.pac_file is not None:
+            default_plugins.append((PLUGIN_PAC_FILE, True))
+        return default_plugins
+
+    @staticmethod
+    def is_py3() -> bool:
+        """Exists only to avoid mocking sys.version_info in tests."""
+        return sys.version_info[0] == 3
+
+    @staticmethod
+    def setup_logger(
+            log_file: Optional[str] = DEFAULT_LOG_FILE,
+            log_level: str = DEFAULT_LOG_LEVEL,
+            log_format: str = DEFAULT_LOG_FORMAT) -> None:
+        ll = getattr(
+            logging,
+            {'D': 'DEBUG',
+             'I': 'INFO',
+             'W': 'WARNING',
+             'E': 'ERROR',
+             'C': 'CRITICAL'}[log_level.upper()[0]])
+        if log_file:
+            logging.basicConfig(
+                filename=log_file,
+                filemode='a',
+                level=ll,
+                format=log_format)
+        else:
+            logging.basicConfig(level=ll, format=log_format)
+
+    @staticmethod
+    def set_open_file_limit(soft_limit: int) -> None:
+        """Configure open file description soft limit on supported OS."""
+        if os.name != 'nt':  # resource module not available on Windows OS
+            curr_soft_limit, curr_hard_limit = resource.getrlimit(
+                resource.RLIMIT_NOFILE)
+            if curr_soft_limit < soft_limit < curr_hard_limit:
+                resource.setrlimit(
+                    resource.RLIMIT_NOFILE, (soft_limit, curr_hard_limit))
+                logger.debug(
+                    'Open file soft limit set to %d', soft_limit)
 
 
 @contextlib.contextmanager
