+++ tests/test_set_open_file_limit.py
@@ -12,7 +12,7 @@ import os
 import unittest
 from unittest import mock
 
-from proxy.common.flags import Flags
+from proxy.proxy import Proxy
 
 if os.name != 'nt':
     import resource
@@ -29,7 +29,7 @@ class TestSetOpenFileLimit(unittest.TestCase):
             self,
             mock_set_rlimit: mock.Mock,
             mock_get_rlimit: mock.Mock) -> None:
-        Flags.set_open_file_limit(256)
+        Proxy.set_open_file_limit(256)
         mock_get_rlimit.assert_called_with(resource.RLIMIT_NOFILE)
         mock_set_rlimit.assert_called_with(resource.RLIMIT_NOFILE, (256, 1024))
 
@@ -39,7 +39,7 @@ class TestSetOpenFileLimit(unittest.TestCase):
             self,
             mock_set_rlimit: mock.Mock,
             mock_get_rlimit: mock.Mock) -> None:
-        Flags.set_open_file_limit(256)
+        Proxy.set_open_file_limit(256)
         mock_get_rlimit.assert_called_with(resource.RLIMIT_NOFILE)
         mock_set_rlimit.assert_not_called()
 
@@ -49,6 +49,6 @@ class TestSetOpenFileLimit(unittest.TestCase):
             self,
             mock_set_rlimit: mock.Mock,
             mock_get_rlimit: mock.Mock) -> None:
-        Flags.set_open_file_limit(1024)
+        Proxy.set_open_file_limit(1024)
         mock_get_rlimit.assert_called_with(resource.RLIMIT_NOFILE)
         mock_set_rlimit.assert_not_called()
